requirements.txt
numpy==1.15.4
Pillow==6.2.0
pyopencl==2018.1.1
